kevin v7

+++ AparatosElectronicos/src/main/java/uptc/frw/aparatoselectronicos/jpa/repository/RepairComponentRepository.java
@@ -3,5 +3,5 @@ package uptc.frw.aparatoselectronicos.jpa.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uptc.frw.aparatoselectronicos.jpa.entity.RepairComponent;
 
-public interface RepairComponentRepository extends JpaRepository<RepairComponent, Integer> {
+public interface RepairComponentRepository extends JpaRepository<RepairComponent, Long> {
 }
