+++ AparatosElectronicos/src/main/java/uptc/frw/aparatoselectronicos/jpa/repository/RepairRepository.java
@@ -3,5 +3,5 @@ package uptc.frw.aparatoselectronicos.jpa.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uptc.frw.aparatoselectronicos.jpa.entity.Repair;
 
-public interface RepairRepository extends JpaRepository<Repair, Integer> {
+public interface RepairRepository extends JpaRepository<Repair, Long> {
 }
